moved Links classes to separate files
